Creation du module : creation_produit, Ajout de la bilbiotheque 'json' pour gerer l'enregistrement des produits dans un fichier '.json'
