+++ main.py
@@ -1,8 +1,4 @@
-produit = []
-num_art = 0
-prix_art = ''
-qte_art = ''
-article_existe = False
+import creation_produit as creation
 
 
 def menu_principal():
@@ -18,52 +14,8 @@ def menu_principal():
     return input("Choisissez une option: ")
 
 
-def testdecimal(valeur):
-    try:
-        float(valeur)
-        return True
-    except:
-        return False
 
-def validation_prix_qte(prix, qte):
-    global prix_art
-    global qte_art
-    while not testdecimal(prix):
-        print("\nErreur de saisie, la valeur du prix doit etre un chiffre svp !")
-        prix = input(f"""Prix unitaire
-==> """)
-        prix_art = prix
 
-    while not qte.isdigit():
-        print("\nErreur de saisie, la valeur de la quantité doit etre un entier svp !")
-        qte = input(f"""Quantité
-==> """)
-        qte_art = qte
-
-def ajout_produit():
-    global article_existe
-    global prix_art
-    global qte_art
-    global num_art
-    num_art += 1
-    print("AJOUT DE ARTICLES")
-    print("==================")
-    nom_art = input(f"""Nom
-==> """)
-    prix_art = input(f"""Prix unitaire
-==> """)
-    qte_art = input(f"""Quantité
-==> """)
-    validation_prix_qte(prix_art, qte_art)
-
-    for i in produit:
-        if nom_art in list(i.values()):
-            print("Ce produit existe déjà")
-            article_existe = 1
-            break
-    if not article_existe:
-        produit.append({'Designation': nom_art, 'Numero': num_art, 'Prix': float(prix_art), 'Quantité': int(qte_art)})
-        return produit
 
 def main():
     # charger_donnees()
@@ -71,7 +23,7 @@ def main():
         choix = menu_principal()
         if choix == '1':
             # print("interface_ajout_produit")
-            print(ajout_produit())
+            print(creation.ajout_produit())
         elif choix == '2':
             # interface_affichage_produits()
             print("interface_affichage_produits")
